Fix try again button issues
Change button visibility for weapons buttons and try again button
Based on conidiotion that either player or computer score is ==5. If ==5
then weapon buttons are no longer visible and the try again button
becomes visible. This is a workaround to prevent the score from
growing past 5.
When try again is pressed weapon choice visibility is now visible,
and try again button is now hidden.